masterstation/requirements.txt
streamlit
psutil
py-cpuinfo
requests
black
pylint
bandit
python-dotenv
